archivist_agent/agent/gemini_rag/requirements.txt
# Dependencias simples para herramientas RAG de Google ADK
aigency
# Procesamiento de documentos
PyMuPDF==1.26.4
numpy==2.3.3
scipy==1.16.2

# Environment variables
python-dotenv

# Embeddings locales (versiones compatibles)
google-generativeai

# Base de datos vectorial
chromadb>=1.1.0